refactor(shared): US-ADJ-2.7 — EventStorePort + SQLiteEventStore a shared/, composition root P-08 en app.py
- Mueve EventStorePort → shared/domain/ports/ (fuente canónica cross-BC)
- Mueve SQLiteEventStore → shared/infrastructure/event_store/ (fuente canónica)
- competencia/domain/ports/ y competencia/infrastructure/event_store/ re-exportan desde shared
- resultados/ importa EventStorePort desde shared.domain.ports
- resultados/api/router.py importa SQLiteEventStore desde shared.infrastructure
- Elimina imports cross-BC de resultados/ en competencia/api/router.py (B-03)
- Elimina función get_on_finalizada_callback (código muerto) del router
- build_on_finalizada_callback vive ahora en src/app.py (composition root)
- 481 tests pasando, 0 regresiones
