+++ src/resultados/infrastructure/repositories/resultados_competencia_adapter.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from uuid import UUID
 
 from competencia.domain.aggregates.performance import Performance
-from competencia.domain.ports.event_store_port import EventStorePort
+from shared.domain.ports.event_store_port import EventStorePort
 from shared.domain.value_objects.disciplina import Disciplina
 from competencia.domain.value_objects.estado_performance import EstadoPerformance
 from resultados.domain.ports.resultados_competencia_port import (
