+++ src/shared/domain/ports/__init__.py
@@ -1,4 +1,4 @@
-"""Re-export de EventStorePort desde shared.domain (fuente canónica)."""
+"""Puertos compartidos entre Bounded Contexts."""
 from shared.domain.ports.event_store_port import EventStorePort
 
 __all__ = ["EventStorePort"]
