+++ src/shared/infrastructure/__init__.py
@@ -1,0 +1,1 @@
+"""Infraestructura compartida entre Bounded Contexts."""
