+++ src/shared/infrastructure/event_store/__init__.py
@@ -1,0 +1,1 @@
+"""Event Store compartido."""
